requirements.txt
pytest==8.0.0
selenium==4.17.2
allure-pytest==2.13.5
webdriver-manager==4.0.1
requests~=2.32.3
jsonschema=4.23.0